Port driver to PCI-rework interfaces.
Quick and dirty port to the PCI-rework interfaces.  Test on a V2100
cards.  This is the only Rendition hardware to which I have access.
V1000 and / or V2200 may be broken, but I doubt it.

Comments in the source claim that acceleration is buggy, and it is
disabled by default.  I *did not* do any testing with acceleration.